[core] Rename `callback` to `Event`
