+++ packages/expo-camera/android/src/main/java/expo/modules/camera/ExpoCameraView.kt
@@ -30,7 +30,6 @@ import expo.modules.interfaces.facedetector.FaceDetectorInterface
 import expo.modules.interfaces.facedetector.FaceDetectorProviderInterface
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.Promise
-import expo.modules.kotlin.callbacks.callback
 import expo.modules.kotlin.views.ExpoView
 import java.io.File
 import java.io.IOException
@@ -42,6 +41,7 @@ import expo.modules.camera.utils.mapY
 import kotlin.math.roundToInt
 import android.view.WindowManager
 import expo.modules.interfaces.barcodescanner.BarCodeScannerResult.BoundingBox
+import expo.modules.kotlin.viewevent.Event
 
 class ExpoCameraView(
   context: Context,
@@ -61,9 +61,9 @@ class ExpoCameraView(
   private var isPaused = false
   private var isNew = true
 
-  private val onCameraReady by callback<Unit>()
-  private val onMountError by callback<CameraMountErrorEvent>()
-  private val onBarCodeScanned by callback<BarCodeScannedEvent>(
+  private val onCameraReady by Event<Unit>()
+  private val onMountError by Event<CameraMountErrorEvent>()
+  private val onBarCodeScanned by Event<BarCodeScannedEvent>(
     /**
      * We want every distinct barcode to be reported to the JS listener.
      * If we return some static value as a coalescing key there may be two barcode events
@@ -73,15 +73,15 @@ class ExpoCameraView(
      */
     coalescingKey = { event -> (event.data.hashCode() % Short.MAX_VALUE).toShort() }
   )
-  private val onFacesDetected by callback<FacesDetectedEvent>(
+  private val onFacesDetected by Event<FacesDetectedEvent>(
     /**
      * Should events about detected faces coalesce, the best strategy will be
      * to ensure that events with different faces count are always being transmitted.
      */
     coalescingKey = { event -> (event.faces.size % Short.MAX_VALUE).toShort() }
   )
-  private val onFaceDetectionError by callback<FaceDetectionErrorEvent>()
-  private val onPictureSaved by callback<PictureSavedEvent>(
+  private val onFaceDetectionError by Event<FaceDetectionErrorEvent>()
+  private val onPictureSaved by Event<PictureSavedEvent>(
     coalescingKey = { event ->
       val uriHash = event.data.getString("uri")?.hashCode() ?: -1
       (uriHash % Short.MAX_VALUE).toShort()
