--- packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/callbacks/Callback.kt
@@ -1,5 +1,0 @@
-package expo.modules.kotlin.callbacks
-
-fun interface Callback<T> {
-  operator fun invoke(arg: T)
-}
