+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/viewevent/ViewEvent.kt
@@ -1,15 +1,16 @@
-package expo.modules.kotlin.callbacks
+package expo.modules.kotlin.viewevent
 
 import android.view.View
 import com.facebook.react.bridge.ReactContext
 import com.facebook.react.bridge.WritableMap
 import expo.modules.adapters.react.NativeModulesProxy
+import expo.modules.kotlin.callbacks.Callback
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.types.JSTypeConverter
 import expo.modules.kotlin.types.putGeneric
 import kotlin.reflect.KType
 
-class ViewCallback<T>(
+class ViewEvent<T>(
   private val name: String,
   private val type: KType,
   private val view: View,
