+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/viewevent/ViewEventDelegate.kt
@@ -1,8 +1,9 @@
 @file:OptIn(ExperimentalStdlibApi::class)
 
-package expo.modules.kotlin.callbacks
+package expo.modules.kotlin.viewevent
 
 import android.view.View
+import expo.modules.kotlin.callbacks.Callback
 import java.lang.ref.WeakReference
 import kotlin.reflect.KProperty
 import kotlin.reflect.KType
@@ -10,7 +11,7 @@ import kotlin.reflect.typeOf
 
 typealias CoalescingKey<T> = (event: T) -> Short
 
-class ViewCallbackDelegate<T>(
+class ViewEventDelegate<T>(
   private val type: KType,
   view: View,
   private val coalescingKey: CoalescingKey<T>?
@@ -25,7 +26,7 @@ class ViewCallbackDelegate<T>(
 
     val view = viewHolder.get()
       ?: throw IllegalStateException("Can't send an event from the view that is deallocated.")
-    return ViewCallback(property.name, type, view, coalescingKey)
+    return ViewEvent(property.name, type, view, coalescingKey)
   }
 }
 
@@ -34,6 +35,11 @@ class ViewCallbackDelegate<T>(
  * @param coalescingKey a key generator used to determine which other events of this type this event can be coalesced with.
  * For example, touch move events should only be coalesced within a single gesture so a coalescing key there would be the unique gesture id.
  */
-inline fun <reified T> View.callback(noinline coalescingKey: CoalescingKey<T>? = null): ViewCallbackDelegate<T> {
-  return ViewCallbackDelegate(typeOf<T>(), this, coalescingKey)
+inline fun <reified T> View.Event(noinline coalescingKey: CoalescingKey<T>? = null): ViewEventDelegate<T> {
+  return ViewEventDelegate(typeOf<T>(), this, coalescingKey)
+}
+
+@JvmName("MapEvent")
+fun View.Event(coalescingKey: CoalescingKey<Map<String, Any>>? = null): ViewEventDelegate<Map<String, Any>> {
+  return ViewEventDelegate(typeOf<Map<String, Any>>(), this, coalescingKey)
 }
